Move download code into rust_utils::download
